requirements.txt
anthropic>=0.50.0
google-cloud-bigquery>=3.0.0
